taobao_mcp/pyproject.toml
[project]
name = "taobao-mcp-server"
version = "1.0.0"
description = "Model Context Protocol server for Taobao/Tmall product scraping"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "playwright>=1.40.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
